switching over to browserify
