--- component.json
@@ -1,16 +1,0 @@
-{
-  "name": "hone-embed",
-  "repo": "honeinc/hone-embed",
-  "description": "A simple tool, to work with Hones postmessage from embeded Hone's",
-  "version": "0.0.3",
-  "keywords": [],
-  "dependencies": {
-    "component/emitter" : "master"
-  },
-  "development": {},
-  "license": "MIT",
-  "main": "index.js",
-  "scripts": [
-    "index.js"
-  ]
-}
