+++ examples/dynamic-src.html
@@ -1,6 +1,6 @@
 <!doctype html>
 <iframe width="600px" height="800px" frameborder="0" data-hone="5407769cfae50de35b93a1c2" data-resize="true">You Dont Support Iframes</iframe>
-<script src="../build/embed.js" defer></script>
+<script src="../build/hone-embed.js" defer></script>
 <script>
 onHoneReady = function ( hone ) {
     hone.init({
